zephyr push

+++ zlog.h
@@ -8,7 +8,7 @@
 #include <iostream>
 #include <fstream>
 #include "zstring.h"
-#include "asar.h"
+//#include "asar.h"
 #include "zthread.h"
 #include "zmutex.h"
 #include "zpath.h"
@@ -61,7 +61,7 @@ public:
     ZLog &log(ZString logtext);
     ZLog &operator<<(ZString text);
     ZLog &operator<<(std::string text);
-    ZLog &operator<<(int text);
+    ZLog &operator<<(long int text);
     ZLog &operator<<(const char *text);
     ZLog &operator<<(char *text);
     ZLog &operator<<(ZPath text);
